Do not unnecessary scroll, attach to bottom if possible
Instead of always attaching an element to the top of its container, it
now changes the scroll position based on its clipping point. That is,
if clipped on top, it scrolls down a bit to get the top visible.
Similarly for the bottom. If there is no clip, do nothing.